# openapi.yaml
openapi: 3.1.0
info:
  title: CORTEX — Sovereign Memory Engine
  description: Enterprise API for persistent AI memory, semantic search, and multi-agent
    swarm coordination.
  version: 4.3.0
  contact:
    name: borjamoskv.com
    email: borjamoskv@gmail.com
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
paths:
...
  /v1/ask:
    post:
      tags:
      - ask
      summary: Ask Cortex
      description: 'RAG endpoint: search → synthesize → answer.


        Searches CORTEX memory for relevant facts, then uses the configured

        LLM to synthesize an answer grounded in those facts.


        Returns 503 if no LLM provider is configured.'
      operationId: ask_cortex_v1_ask_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Bearer <api-key>
          title: Authorization
        description: Bearer <api-key>
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AskRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/llm/status:
    get:
      tags:
      - ask
      summary: Llm Status
      description: Check LLM provider status and list supported providers.
      operationId: llm_status_v1_llm_status_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Bearer <api-key>
          title: Authorization
        description: Bearer <api-key>
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AskRequest:
      properties:
        query:
          type: string
          maxLength: 4096
          minLength: 1
          title: Query
          description: Natural language question
        project:
          anyOf:
          - type: string
          - type: 'null'
          title: Project
          description: Filter by project (optional)
        k:
          type: integer
          maximum: 50.0
          minimum: 1.0
          title: K
          description: Number of facts to retrieve
          default: 10
        temperature:
          type: number
          maximum: 2.0
          minimum: 0.0
          title: Temperature
          description: LLM sampling temperature
          default: 0.3
        max_tokens:
          type: integer
          maximum: 8192.0
          minimum: 64.0
          title: Max Tokens
          description: Max response tokens
          default: 2048
        system_prompt:
          anyOf:
          - type: string
          - type: 'null'
          title: System Prompt
          description: Override system prompt (optional)
      type: object
      required:
      - query
      title: AskRequest
      description: 'RAG query: search CORTEX memory and synthesize an answer.'
    AskResponse:
      properties:
        answer:
          type: string
          title: Answer
        sources:
          items:
            $ref: '#/components/schemas/AskSource'
          type: array
          title: Sources
        model:
          type: string
          title: Model
        provider:
          type: string
          title: Provider
        facts_found:
          type: integer
          title: Facts Found
      type: object
      required:
      - answer
      - sources
      - model
      - provider
      - facts_found
      title: AskResponse
      description: RAG response with answer and sources.
    AskSource:
      properties:
        fact_id:
          type: integer
          title: Fact Id
        content:
          type: string
          title: Content
        score:
          type: number
          title: Score
        project:
          type: string
          title: Project
      type: object
      required:
      - fact_id
      - content
      - score
      - project
      title: AskSource
      description: A source fact that contributed to the answer.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LLMStatusResponse:
      properties:
        available:
          type: boolean
          title: Available
        provider:
          type: string
          title: Provider
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        supported_providers:
          items:
            type: string
          type: array
          title: Supported Providers
      type: object
      required:
      - available
      - provider
      - supported_providers
      title: LLMStatusResponse
      description: LLM provider status.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...